Seperate out into their own lines

+++ builder_scripts/build_snapshots_looped.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 #
 # pfSense snapshot building system
-# (C)2007, 2008 Scott Ullrich
+# (C)2007, 2008, 2009 Scott Ullrich
 # All rights reserved
 #
 # Redistribution and use in source and binary forms, with or without
@@ -16,9 +16,33 @@
 #    documentation and/or other materials provided with the distribution.
 #
 
-rm -f /tmp/pfSense_do_not_build_pfPorts
+# Handle command line arguments
+while test "$1" != "" ; do
+	case $1 in
+		--noports|-n)
+		echo "$2"
+		NO_PORTS=yo
+		shift
+	;;
+	esac
+	shift
+done
 
+# Main builder loop
+COUNTER=0
 while [ /bin/true ]; do
+	# We can disable ports builds
+	if [ "$NO_PORTS" = "" ]; then
+		echo ">>> Not building pfPorts at all during this snapshot builder looped run..."
+		touch /tmp/pfSense_do_not_build_pfPorts
+	else
+		if [ "$COUNTER" -gt 0 ]; then 
+			echo ">>> Previous snapshot runs deteceted, not building pfPorts again..."
+			touch /tmp/pfSense_do_not_build_pfPorts
+		else
+			rm -f /tmp/pfSense_do_not_build_pfPorts
+		fi
+	fi
 	./build_snapshots.sh
 	# Grab a random value and sleep
 	value=`od -A n -d -N2 /dev/random | awk '{ print $1 }'`
@@ -26,5 +50,9 @@ while [ /bin/true ]; do
 	echo
 	echo ">>> Sleeping for $value in between snapshot builder runs"
 	echo
+	# Count some sheep.
 	sleep $value
+	COUNTER=`expr $COUNTER + 1`
+	echo ">>> Starting builder run #${COUNTER}..."
+	echo
 done
